Tidy up the modification function and move into utils
